fixed cargo/carrier vehicle ops
